Improve testing setup
* Test via Mocha
* Configure WebPack to spit out a UMD module
