+++ memo_wasm/package.json
@@ -3,10 +3,12 @@
   "version": "0.1.0",
   "main": "dist/index.js",
   "scripts": {
-    "install": "script/build"
+    "install": "script/build",
+    "test": "mocha --ui=tdd test"
   },
   "license": "MIT",
   "devDependencies": {
+    "mocha": "^5.2.0",
     "webpack": "^4.6.0",
     "webpack-cli": "^2.0.15"
   }
