+++ memo_wasm/script/build
@@ -20,4 +20,4 @@ rm -rf dist
 mkdir -p dist
 CARGO_INCREMENTAL=0 RUSTFLAGS="-C debuginfo=0 -C opt-level=s -C lto -C panic=abort" cargo build --release --target wasm32-unknown-unknown
 wasm-bindgen ../target/wasm32-unknown-unknown/release/memo_wasm.wasm --out-dir dist
-node_modules/.bin/webpack --target=node --mode=development lib/index.js --output-filename=index.js
+node_modules/.bin/webpack --mode=development
