+++ memo_wasm/webpack.config.js
@@ -1,0 +1,12 @@
+var path = require("path");
+
+module.exports = {
+  entry: "./src/index.js",
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "index.js",
+    library: "memo",
+    libraryTarget: "umd"
+  },
+  target: "node"
+};
